KWSys: Fix conversion warning in testProcess.c
Add an explicit cast to avoid a warning about 'int' -> 'unsigned int'
conversion sign change.  We know our values are positive anyway.
Fix the warning here while we wait for upstream KWSys to address it.

+++ Source/kwsys/testProcess.c
@@ -394,7 +394,7 @@ static int runChild2(kwsysProcess* kp,
 
   if(interruptDelay)
     {
-    testProcess_sleep(interruptDelay);
+    testProcess_sleep((unsigned int)interruptDelay);
     kwsysProcess_Interrupt(kp);
     }
